+++ examples/resources/sonatypeiq_config_crowd/resource.tf
@@ -1,0 +1,5 @@
+resource "sonatypeiq_config_crowd" "crowd" {
+  server_url           = "http://something2"
+  application_name     = "name"
+  application_password = "something"
+}